backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.0
httpx>=0.28
python-dotenv>=1.1
google-genai>=1.0
